Have the gdb watcher try to match exit code better
You can't use exit() to exit in the same way as if you had
died with a signal; we tried to do some exit-code munging
but we did it one way and our tester script expected another
way.  we could resolve that, or use the approach here: try
to die with the same signal by sending it to ourselves.