Allow passing NULL for empty kwargs
This is a convention that cpython uses that is a good
performance improvement -- unlike tuples, dicts are mutable,
so we currently have to create a new dict for every call
into a kwargs-taking function even if there are no keywords
to pass.

Have to modify the receiving ends to allow NULL kwargs.  This
includes the ast interpreter (which I think will automatically get
the bjit), the llvm irgenerator, and as many runtime functions as
I can find.