section ordering hack: use the PGO-determined ordering
Rather than do anything smart, just copy the ordering that the
pgo build used.  To keep the file reasonable, only emit
entries for functions that appear in the perf results (cuts from
>1M entries to ~2k).

I tried getting the hhvm "hfsort" to work, but ran into some
fixable-but-annoying issues getting it integrated (ex need to bring
in some other libraries).  Let's just try this for now; the goal
isn't to get the absolute best performance anyway, since for that
we'll use PGO and ignore the section ordering file.