Build fix: I think this flag confuses clang
Haven't looked into it that much, but my guess is that it
was introduced in some newer version of clang than what
some people are using.  It was added for the use of the
clang-pgo build, but we ended up dropping that so it should
be ok to drop this flag.