Try to fix libunwind build issue
My guess is that there's a build dependency issue, where
we try building files that include libunwind.h before we
run the rules that generate it.  Not sure why we weren't running
into that before.  But the normal c++ files in src/ correctly
depend on libunwind, but the ones in src/runtime/inline don't
seem to.

For good measure, don't include libunwind.h in our header files,
and just use forward-declaring.