Get assembly comments working for the bjit
First issue is that the assembly logging previously worked
on the assembler buffer, which is reasonable but has issues with the bjit
which will update the assembly post-relocation (ie by initializing the patchpoints).

Also add some extra flags to separate commenting for rewriter-level comments
("this is the beginning of a setattr") and the new bjit ones ("this is
a set local").