Don't create Register decref-info entries
They might be ok if we allowed allocating callee-save registers, but
since we don't, they mean we are asking the unwinder to restore a
clobbered register.

This was happening in some places where we didn't call setupCall()
before calling a function.  This was previously ok since those cases would
always throw an exception and nothing would want the values of the
then-clobbered registers.