Don't have attrwrapper.get convert to dict-backed
The issue with django was that it called object.__dict__.__get__(unicode),
which would convert object to dict-backed, and then our shutdown logic doesn't
support that case.

The shutdown logic still doesn't support that case, and there are other ways
to convert object's __dict__ to being dict-backed, so this commit isn't a full
solution, but it seems like something we should do anyway.