Workaround for the "terminator in middle of basic block" issue
The problem is that we emit an llvm "unreachable" instruction, and then
continue to emit other code, which fails the verifier.

endBlock(DEAD) is supposed to be the right way to handle that, but there
is some more work that would need to be done there to get that working
properly.

So just do the easy thing for now -- create a new BB so that it's ok to
emit more code.