Bump build time to 60 minutes
Somehow this PR is taking longer to build, I assume from somehow missing
the build cache.

Assuming that it's a caching issue, bump up the time to satisfy un-cached builds.
If build times continue to increase we will run into this problem again (can only
regenerate the cache with a successful build but a successful build needs the cache),
but this will let us defer it a bit longer :)  Travis-CI builds are max 120 minutes.