Cleanup: iterating a VRegMap skips default entries
The VRegMap doesn't contain explicit "is-set" bits, so I originally
just had iteration return all values.

But all the uses of it ended up needing to have a default value to represent
"not set in map" and then iterating would explicitly check if the value was
that and then would skip it.

Now that behavior is moved into VRegMap itself.