Support empty lines on the repl
They end up generating "pass" statements with a lineno of 0, which
trips an assert later on.  This commit just sets them to have a lineno
of 1.

I'm not sure how to test this, since piping into stdin is supposed to
be treated as a file (not as the repl).  Though, we get that wrong right now.