Undo all the places we changed installed paths
We used to have a installed directory structure that mirrors
our source structure.  This required changing quite a few pieces of CPython
code to look in the new places (virtualenv, distutils, getpath.c, sysconfig).
So, undo all those changes, and change our CMake configuration to install into
the paths that CPython expects.

Note: "the paths that CPython expects" is platform-dependent.  The paths I added
here are based on the "posix" style that it supports.  The CPython code has platform-specific
code to look in different places, and I guess the build system must have platform-specific
knowledge of where to put it as well, but our CMake system will always put it in the linux variant.