Turn off implicit-function check in release mode
I'm still hoping that we can keep this on in some way, because it continues to find issues.
So let's try keeping it on for debug builds.

I'm not sure if this is weakening the check enough (there may be other cases like scipy),
but it's sounding like we need to have this check turned off in our releases regardless,
such as issue #1344.