Properly throw ImportErrors on dlopen failure
I think it'd be nicer to just abort() like we do here, but
cffi is actually testing this behavior (that loading an invalid
library causes an ImportError).

This is being exposed now because before, the library would fail
to compile due to -Werror=implicit-function-declaration and not even
get to the loading step.